wstest: alternate home, don't test ASAN
Use `alternate-home.sh` when running the test suite, to fully rule out
any pollution from my dot files which could be tainting the test
results.

Check if Wireshark was compiled using ASAN, and skip testing entirely if
so. The ASAN output breaks the majority of test cases. This still lets
me use `wstest` as a short wrapper for rebuilding Wireshark.